requirements.txt
spacy
scispacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz
torch
accelerate
peft
datasets
scikit-learn
neo4j
biopython
flwr[simulation]
networkx
torch-geometric
numpy
rank_bm25
sentence-transformers
faiss-cpu
pysbd
joblib
conllu
nmslib-metabrainz
pydantic<2.0
typing-extensions